fix: separate CLI main function to prevent auto-execution on import
- Create cli/main.ts with cliMain function (no auto-execution)
- Update cli/bin.ts to import and execute cliMain
- Export cliMain from index.node.ts without side effects

This allows mulmocast-easy to import cliMain without triggering CLI execution.
