+++ src/index.node.ts
@@ -10,4 +10,4 @@ export * from "./methods/index.js";
 export * from "./agents/index.js";
 
 // CLI entry point for programmatic usage (e.g., mulmocast-easy)
-export { main as cliMain } from "./cli/bin.js";
+export { cliMain } from "./cli/main.js";
